requirements.txt
aiohttp~=3.13.2
aiogram>=3.25.0
kurigram==2.2.19
pysocks~=1.7.1
pillow~=12.0.0
psutil~=7.2.1
pymongo>=4.15.5
pytgcrypto>=1.2.11
py-tgcalls~=2.2.10
py-yt-search~=0.5.5
python-dotenv~=1.2.1
yt-dlp[default]>=2026.03.17
